Modify get_smart.sh to append SAS controller information for each disk
Each disk is connected to a certain SAS controller,
and MAG-244 is depending on this information to fill the SPMF
simple storage schema.

The SAS controller information will further be decoded in
on-tasks (command-parser.js) thus it will be added to the
smart catalog.

This commit also contains an minor fix to minor issue on checking
duplicate disks(line 59)